Stop needlessly sorting the hooks list while OOG
Profiling shows that the bulk of our time out of game is spent in
Hooks.sort(),
which is completely useless since the Z orders won't get changed OOG,
and the
hooks aren't visible OOG anyway. Skipping the sort in this case can
result in
a visible improvement in apparent "lagginess".